# clvImob-main/server/swagger.json
{
    "openapi" : "3.0.0",
    "info" : {
        "Title": "CLVImob API",
        "description": "Consult API end-points to the site",
        "contact": "cheljose@gmail.com",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/",
            "description": "Local server"
        }
    ],
    "paths": {
...
    "/{id}": {
        "get": {
            "summary": "Obtém um recurso específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "findById",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Recurso obtido com sucesso"}}
        },
        "delete": {
            "summary": "Deleta um recurso específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "deleteContract",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Recurso deletado com sucesso"}}
        },
        "put": {
            "summary": "Atualiza um recurso específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "updateDocumento",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Recurso atualizado com sucesso"}}
        },
        "patch": {
            "summary": "Atualiza parcialmente um recurso específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "updateImob",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Recurso atualizado parcialmente com sucesso"}}
        }
    },
    "/top": {
        "get": {
            "summary": "Obtém os principais imóveis",
            "security": [{"authMiddleware": []}],
            "operationId": "topImob",
            "responses": {"200": {"description": "Imóveis obtidos com sucesso"}}
        }
    },
    "/busca/cidade": {
        "get": {
            "summary": "Busca recursos por cidade",
            "security": [{"authMiddleware": []}],
            "operationId": "searchByCidade",
            "responses": {"200": {"description": "Recursos encontrados com sucesso"}}
        }
    },
    "/busca/estado": {
        "get": {
            "summary": "Busca recursos por estado",
            "security": [{"authMiddleware": []}],
            "operationId": "searchByEstado",
            "responses": {"200": {"description": "Recursos encontrados com sucesso"}}
        }
    },
    "/busca/endereco": {
        "get": {
            "summary": "Busca recursos por endereço",
            "security": [{"authMiddleware": []}],
            "operationId": "searchByEndereco",
            "responses": {"200": {"description": "Recursos encontrados com sucesso"}}
        }
    },
    "/busca/cep": {
        "get": {
            "summary": "Busca recursos por CEP",
            "security": [{"authMiddleware": []}],
            "operationId": "searchByCep",
            "responses": {"200": {"description": "Recursos encontrados com sucesso"}}
        }
    },
    "/documentos": {
        "post": {
            "summary": "Cria um novo documento",
            "security": [{"authMiddleware": []}],
            "operationId": "createDocumento",
            "responses": {"200": {"description": "Documento criado com sucesso"}}
        },
        "get": {
            "summary": "Obtém todos os documentos",
            "security": [{"authMiddleware": []}],
            "operationId": "getAllDocumentos",
            "responses": {"200": {"description": "Documentos obtidos com sucesso"}}
        }
    },
    "/documentos/{id}": {
        "get": {
            "summary": "Obtém um documento específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "getDocumentoById",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Documento obtido com sucesso"}}
        },
        "delete": {
            "summary": "Deleta um documento específico por ID",
            "security": [{"authMiddleware": []}],
            "operationId": "deleteDocumento",
            "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
            "responses": {"200": {"description": "Documento deletado com sucesso"}}
        }
    },
...
    },
...
}